+++ build_runner_core/test/asset/cache_test.dart
@@ -12,14 +12,12 @@ void main() {
   var missingTxt = AssetId('a', 'missing.txt');
   var fooContent = 'bar';
   var fooutf8Bytes = decodedMatches('bar');
-  var assets = <AssetId, dynamic>{
-    fooTxt: 'bar',
-  };
   late InMemoryRunnerAssetReaderWriter delegate;
   late CachingAssetReader reader;
 
   setUp(() {
-    delegate = InMemoryRunnerAssetReaderWriter(sourceAssets: assets);
+    delegate = InMemoryRunnerAssetReaderWriter()
+      ..filesystem.writeAsStringSync(fooTxt, 'bar');
     reader = CachingAssetReader(delegate);
   });
 
@@ -27,104 +25,104 @@ void main() {
     test('should read from the delegate', () async {
       expect(await reader.canRead(fooTxt), isTrue);
       expect(await reader.canRead(missingTxt), isFalse);
-      expect(delegate.assetsRead, {fooTxt, missingTxt});
+      expect(delegate.inputTracker.assetsRead, {fooTxt, missingTxt});
     });
 
     test('should not re-read from the delegate', () async {
       expect(await reader.canRead(fooTxt), isTrue);
-      delegate.assetsRead.clear();
+      delegate.inputTracker.assetsRead.clear();
       expect(await reader.canRead(fooTxt), isTrue);
-      expect(delegate.assetsRead, isEmpty);
+      expect(delegate.inputTracker.assetsRead, isEmpty);
     });
 
     test('can be invalidated with invalidate', () async {
       expect(await reader.canRead(fooTxt), isTrue);
-      delegate.assetsRead.clear();
-      expect(delegate.assetsRead, isEmpty);
+      delegate.inputTracker.assetsRead.clear();
+      expect(delegate.inputTracker.assetsRead, isEmpty);
 
       reader.invalidate([fooTxt]);
       expect(await reader.canRead(fooTxt), isTrue);
-      expect(delegate.assetsRead, [fooTxt]);
+      expect(delegate.inputTracker.assetsRead, [fooTxt]);
     });
   });
 
   group('readAsBytes', () {
     test('should read from the delegate', () async {
       expect(await reader.readAsBytes(fooTxt), fooutf8Bytes);
-      expect(delegate.assetsRead, [fooTxt]);
+      expect(delegate.inputTracker.assetsRead, [fooTxt]);
     });
 
     test('should not re-read from the delegate', () async {
       expect(await reader.readAsBytes(fooTxt), fooutf8Bytes);
-      delegate.assetsRead.clear();
+      delegate.inputTracker.assetsRead.clear();
       expect(await reader.readAsBytes(fooTxt), fooutf8Bytes);
-      expect(delegate.assetsRead, isEmpty);
+      expect(delegate.inputTracker.assetsRead, isEmpty);
     });
 
     test('can be invalidated with invalidate', () async {
       expect(await reader.readAsBytes(fooTxt), fooutf8Bytes);
-      delegate.assetsRead.clear();
-      expect(delegate.assetsRead, isEmpty);
+      delegate.inputTracker.assetsRead.clear();
+      expect(delegate.inputTracker.assetsRead, isEmpty);
 
       reader.invalidate([fooTxt]);
       expect(await reader.readAsBytes(fooTxt), fooutf8Bytes);
-      expect(delegate.assetsRead, [fooTxt]);
+      expect(delegate.inputTracker.assetsRead, [fooTxt]);
     });
 
     test('should not cache bytes during readAsString calls', () async {
       expect(await reader.readAsString(fooTxt), fooContent);
-      expect(delegate.assetsRead, [fooTxt]);
-      delegate.assetsRead.clear();
+      expect(delegate.inputTracker.assetsRead, [fooTxt]);
+      delegate.inputTracker.assetsRead.clear();
 
       expect(await reader.readAsBytes(fooTxt), fooutf8Bytes);
-      expect(delegate.assetsRead, [fooTxt]);
+      expect(delegate.inputTracker.assetsRead, [fooTxt]);
     });
   });
 
   group('readAsString', () {
     test('should read from the delegate', () async {
       expect(await reader.readAsString(fooTxt), fooContent);
-      expect(delegate.assetsRead, [fooTxt]);
+      expect(delegate.inputTracker.assetsRead, [fooTxt]);
     });
 
     test('should not re-read from the delegate', () async {
       expect(await reader.readAsString(fooTxt), fooContent);
-      delegate.assetsRead.clear();
+      delegate.inputTracker.assetsRead.clear();
       expect(await reader.readAsString(fooTxt), fooContent);
-      expect(delegate.assetsRead, isEmpty);
+      expect(delegate.inputTracker.assetsRead, isEmpty);
     });
 
     test('can be invalidated with invalidate', () async {
       expect(await reader.readAsString(fooTxt), fooContent);
-      delegate.assetsRead.clear();
-      expect(delegate.assetsRead, isEmpty);
+      delegate.inputTracker.assetsRead.clear();
+      expect(delegate.inputTracker.assetsRead, isEmpty);
 
       reader.invalidate([fooTxt]);
       expect(await reader.readAsString(fooTxt), fooContent);
-      expect(delegate.assetsRead, [fooTxt]);
+      expect(delegate.inputTracker.assetsRead, [fooTxt]);
     });
 
     test('uses cached bytes if available', () async {
       expect(await reader.readAsBytes(fooTxt), fooutf8Bytes);
-      expect(delegate.assetsRead, [fooTxt]);
-      delegate.assetsRead.clear();
+      expect(delegate.inputTracker.assetsRead, [fooTxt]);
+      delegate.inputTracker.assetsRead.clear();
 
       expect(await reader.readAsString(fooTxt), fooContent);
-      expect(delegate.assetsRead, isEmpty);
+      expect(delegate.inputTracker.assetsRead, isEmpty);
     });
   });
 
   group('digest', () {
     test('should read from the delegate', () async {
       expect(await reader.digest(fooTxt), isNotNull);
-      expect(delegate.assetsRead, [fooTxt]);
+      expect(delegate.inputTracker.assetsRead, [fooTxt]);
     });
 
     test('should re-read from the delegate (no cache)', () async {
       expect(await reader.digest(fooTxt), isNotNull);
-      delegate.assetsRead.clear();
+      delegate.inputTracker.assetsRead.clear();
       expect(await reader.digest(fooTxt), isNotNull);
-      expect(delegate.assetsRead, [fooTxt]);
+      expect(delegate.inputTracker.assetsRead, [fooTxt]);
     });
   });
 }